requirements.txt
amqp==5.2.0
annotated-types==0.7.0
anyio==4.4.0
asgiref==3.8.1
billiard==4.2.0
blinker==1.8.2
boto3==1.34.130
botocore==1.34.130
celery==5.4.0
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
dash==2.17.0
dash-core-components==2.0.0
dash-html-components==2.0.0
dash-player==1.1.0
dash-table==5.0.0
django==4.2.13
django-channels==0.7.0
django-cors-headers==4.3.1
django-environ==0.11.2
django-storages==1.14.3
djangorestframework==3.15.1
dnspython==2.6.1
email-validator==2.1.1
fastapi==0.111.0
fastapi-cli==0.0.4
flask==3.0.3
h11==0.14.0
httpcore==1.0.5
httptools==0.6.1
httpx==0.27.0
idna==3.7
importlib-metadata==7.1.0
itsdangerous==2.2.0
jinja2==3.1.4
jmespath==1.0.1
kombu==5.3.7
markdown-it-py==3.0.0
markupsafe==2.1.5
mdurl==0.1.2
nest-asyncio==1.6.0
oauthlib==3.2.2
orjson==3.10.3
packaging==24.0
plotly==5.22.0
prompt-toolkit==3.0.47
pydantic==2.7.2
pydantic-core==2.18.3
pygments==2.18.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-multipart==0.0.9
python-slugify==8.0.4
pyyaml==6.0.1
redis==5.0.6
requests==2.32.3
requests-oauthlib==2.0.0
retrying==1.3.4
rich==13.7.1
s3transfer==0.10.1
setuptools==70.0.0
shellingham==1.5.4
six==1.16.0
sniffio==1.3.1
sqlparse==0.5.0
starlette==0.37.2
tenacity==8.3.0
text-unidecode==1.3
typer==0.12.3
typing-extensions==4.12.0
tzdata==2024.1
ujson==5.10.0
urllib3==2.2.1
uvicorn==0.30.0
uvloop==0.19.0
vine==5.1.0
watchfiles==0.22.0
wcwidth==0.2.13
websockets==12.0
werkzeug==3.0.3
zipp==3.19.0
